Readme
Dear grader,

Please find attached the reports from the MovieLens project. The R Markdown includes the report with TOC, Introduction, Methodology where the estimated rating is calculated using regularization of the user effect and movie bias on the train and test sets, subsets from the edx dataset. Then, once the model is optimized with proper lambda, the model is evaluated in the edx and validation datasets.

Thank you,
John Estrada